% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_threat_layer_selection.R
\name{add_layer_to_collection}
\alias{add_layer_to_collection}
\title{Add layer to collection}
\usage{
add_layer_to_collection(raster_object, selection, stored_layers, layer_counter)
}
\description{
Add layer to collection
}
