% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceacws_app.R
\name{ceacws_app}
\alias{ceacws_app}
\title{Launch CEACWS Threat Analysis Application}
\usage{
ceacws_app(...)
}
\arguments{
\item{...}{arguments to pass to golem_opts.
See \code{?golem::get_golem_options} for more details.}
}
\value{
A Shiny application object
}
\description{
This function launches the CEACWS (Canadian Environmental Assessment
Canadian Wildlife Service) Shiny application for analyzing threats to
seabirds in the Northwest Atlantic.
}
\examples{
\dontrun{
# Launch the CEACWS application
ceacws_app()
}
}
