% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_threat_layers.R
\name{create_simple_structure}
\alias{create_simple_structure}
\title{Create simple structure for a category}
\usage{
create_simple_structure(df)
}
\arguments{
\item{df}{Data frame with layer information}
}
\value{
List with types, years, months, nighttime, layers
}
\description{
Create simple structure for a category
}
