% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_threat_layer_selection.R
\name{generate_layer_name}
\alias{generate_layer_name}
\title{Generate user-friendly layer name from filters}
\usage{
generate_layer_name(selection)
}
\description{
Generate user-friendly layer name from filters
}
