% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_grid}
\alias{make_grid}
\title{Creates a raster study grid}
\usage{
make_grid(aoi, crs = 4326, cellsize)
}
\arguments{
\item{aoi}{object of class sf or sfc representing the area of interest, or the bounding box of the area to consider. The bounding box should be either a \code{bbox} class object from the \code{sf} package or of the form \code{c(xmin = xmin, ymin = ymin, xmax = xmax, ymax = ymax)}.}

\item{crs}{object of class crs; coordinate reference system of the target of the target grid in case argument x is missing, if x is not missing, its crs is inherited.}

\item{cellsize}{target grid cell size, in the units of the crs selected.}
}
\value{
a tif file and a csv file
}
\description{
This function is used to create a raster study grid that will serve as a template for data extractions
}
\examples{
\dontrun{
cellsize <- 1
crs <- 4326
bb <- c(xmin = -45, ymin = -45, xmax = 45, ymax = 45)
bbox <- sf::st_bbox(bb, crs = crs)
aoi <- sf::st_as_sfc(bbox, crs = crs)
make_grid(bb, crs, cellsize)
make_grid(bbox, crs, cellsize)
make_grid(aoi, crs, cellsize)
}
}
