% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mask_sf}
\alias{mask_sf}
\title{Function to mask stars object with sf polygon}
\usage{
mask_sf(dat, aoi)
}
\arguments{
\item{dat}{stars object to mask using sf polygon}

\item{aoi}{sf polygonused to mask the stars}
}
\description{
Function to mask stars object with sf polygon
}
