% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_threat_layer_selection.R
\name{remove_layer_from_collection}
\alias{remove_layer_from_collection}
\title{Remove layer from collection}
\usage{
remove_layer_from_collection(layer_id, stored_layers)
}
\description{
Remove layer from collection
}
