% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threat_pollution.R
\name{threat_pollution}
\alias{threat_pollution}
\title{Script to generate the threat layer associated with ocean pollution}
\usage{
threat_pollution()
}
\description{
Script to generate the threat layer associated with ocean pollution
}
